feat: add startup service and move MyApp
