+++ test/widget_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:notes_reminder_app/main.dart';
+import 'package:notes_reminder_app/app.dart';
 import 'package:notes_reminder_app/providers/note_provider.dart';
 import 'package:provider/provider.dart';
